Refactor events.py file
