+++ fideslog/sdk/python/events/analytics_event.py
@@ -5,7 +5,7 @@ from typing import Dict, List, Optional
 
 from validators import url as is_valid_url
 
-from .exceptions import InvalidEventError
+from ..exceptions import InvalidEventError
 
 ALLOWED_HTTP_METHODS = [
     "CONNECT",
@@ -120,46 +120,6 @@ class AnalyticsEvent:
                 assert self.status_code > 0 and (
                     self.status_code < 200 or self.status_code > 299
                 ), "An error was provided, but the provided status_code indicates success"
-
-        except AssertionError as err:
-            raise InvalidEventError(str(err)) from None
-
-
-class UserRegistrationEvent:
-    """
-    A discrete event, representing a user action within a fides tool.
-    """
-
-    def __init__(
-        self,
-        email: str,
-        organization: str,
-        registered_at: datetime,
-    ) -> None:
-        """
-        Define a new user registration event to send to the fideslog server.
-
-        :param email: User email
-        :param organization: User organization
-        :param registered_at: The UTC timestamp when the event occurred, in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format. Must include the UTC timezone, and represent a datetime in the past.
-
-        """
-
-        try:
-            assert email is not None and email.find("@") != -1
-            self.email = email
-
-            assert organization is not None
-            self.organization = organization
-
-            assert (
-                registered_at.tzinfo is not None
-                and registered_at.tzinfo == timezone.utc
-            ), "event_created_at must use the UTC timezone"
-            assert registered_at < datetime.now(
-                timezone.utc
-            ), "event_created_at must be in the past"
-            self.registered_at = registered_at
 
         except AssertionError as err:
             raise InvalidEventError(str(err)) from None
