+++ tests/sdk/test_sdk.py
@@ -6,7 +6,7 @@ from typing import Generator
 import pytest
 
 from fideslog.sdk.python.client import AnalyticsClient
-from fideslog.sdk.python.event import AnalyticsEvent
+from fideslog.sdk.python.events.analytics_event import AnalyticsEvent
 
 
 @pytest.fixture()
